+++ theme/template/functions/consts.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Category names.
+ */
+define('NEWS_CATEGORY', 'news');
+define('SPEAKER_SPONSORED', 'sponsored-speakers');
+define('SPEAKER_COMMUNITY', 'community-speakers');
+define('SPONSOR_PLATINUM', 'platinum-sponsors');
+define('SPONSOR_GOLD', 'gold-sponsors');
+define('SPONSOR_SILVER', 'silver-sponsors');
+
+/**
+ * Page names.
+ */
+define('SPEAKERS_PAGE', 'speakers');
+define('SPONSORS_PAGE', 'sponsors');
+
+/**
+ * Custom post types.
+ */
+define('SESSION_POST_TYPE', 'session');