+++ theme/template/functions/theme.php
@@ -1,3 +1,0 @@
-<?php
-// TODO: implement
-add_theme_support('post-thumbnails');